move logic around to satisfy pre-merge hook
